requirements.txt
# BUSI Breast Cancer Detection - Data Preparation & Training Requirements
# Core Data Processing
numpy>=1.20.0
scipy>=1.7.0
Pillow>=9.0.0
scikit-image>=0.18.0
opencv-python>=4.5.0
pandas>=1.3.0

# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Model Architecture
timm>=0.6.0

# Visualization & Analysis
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
pyyaml>=5.4
tqdm>=4.62.0
Pillow>=9.0.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
ipython>=7.0.0
